chore: Algolia 経由の検索実装を撤去する
D1 FTS5 検索 (PR #3444) が本番動作確認済みになったので、Algolia 並走のために残していた gatsby-plugin-algolia / react-instantsearch / モック JSON / GATSBY_SEARCH_PROVIDER 分岐 / 関連環境変数記述を全削除し、検索経路を D1 一本に集約する。Algolia 無料プランのインデックス消失リスクを恒久的に排除し、不要な依存・コードパスを抱え続けないことが目的。
