+++ gatsby-config.ts
@@ -1,4 +1,4 @@
-import { queries, feedOptions, remarkRelatedPostsOptions } from "./src/gatsby/config"
+import { feedOptions, remarkRelatedPostsOptions } from "./src/gatsby/config"
 
 import type { GatsbyConfig } from "gatsby"
 
@@ -112,25 +112,6 @@ const plugins: GatsbyConfig["plugins"] = [
     options: {
       dsn: `https://a52b2817a4214407b72c88d2d8d62ca7@o554110.ingest.sentry.io/5682225`,
       sampleRate: 0.7,
-    },
-  },
-  {
-    resolve: `gatsby-plugin-algolia`,
-    options: {
-      appId: process.env.ALGOLIA_APP_ID,
-      // Use Admin API key without GATSBY_ prefix, so that the key isn't exposed in the application
-      // Tip: use Search API key with GATSBY_ prefix to access the service from within components
-      apiKey: process.env.ALGOLIA_API_KEY,
-      indexName: process.env.ALGOLIA_INDEX_NAME, // for all queries
-      queries,
-      chunkSize: 10000, // default: 1000
-      settings: {
-        // optional, any index settings
-        // Note: by supplying settings, you will overwrite all existing settings on the index
-      },
-      concurrentQueries: false, // default: true
-      skipIndexing: process.env.CF_PAGES_BRANCH !== "master" || process.env.CONTENT_CHANGED === "false", // default: false, useful for e.g. preview deploys or local development
-      continueOnFailure: false, // default: false, don't fail the build if algolia indexing fails
     },
   },
   // this (optional) plugin enables Progressive Web App + Offline functionality
