+++ package.json
@@ -14,7 +14,6 @@
     "dompurify": "3.4.1",
     "dotenv": "17.4.2",
     "gatsby": "5.16.1",
-    "gatsby-plugin-algolia": "1.0.3",
     "gatsby-plugin-feed": "5.16.0",
     "gatsby-plugin-google-analytics": "5.16.0",
     "gatsby-plugin-google-tagmanager": "5.16.0",
@@ -37,7 +36,6 @@
     "prismjs": "1.30.0",
     "react": "19.2.5",
     "react-dom": "19.2.5",
-    "react-instantsearch": "^7.2.0",
     "react-share": "^5.0.0",
     "react-typography": "0.16.23",
     "sass": "1.99.0",
