+++ src/gatsby/config.ts
@@ -1,87 +1,3 @@
-type Data = {
-  data: {
-    postsRemark: {
-      edges: {
-        node: Queries.MarkdownRemark
-      }[]
-    }
-  }
-}
-
-const escapeStringRegexp = (str: string): string => {
-  if (typeof str !== "string") {
-    throw new TypeError("Expected a string")
-  }
-
-  const matchOperatorsRe = /[|\\{}()[\]^$+*?.]/g
-
-  return str.replace(matchOperatorsRe, "\\$&")
-}
-const pagePath = `content/blog`
-
-const algoliaQuery = `
-{
-  postsRemark: allMarkdownRemark(
-    filter: {
-      fileAbsolutePath: { regex: "/${escapeStringRegexp(pagePath)}/" },
-    }
-  ) {
-    edges {
-      node {
-        id
-        frontmatter {
-          date
-          title
-          tags
-          description
-        }
-        fields {
-          slug
-        }
-        internal {
-          contentDigest
-        }
-        rawMarkdownBody
-        timeToRead
-      }
-    }
-  }
-}`
-
-const pageToAlgoliaRecord = ({
-  node: { id, frontmatter, fields, rawMarkdownBody, ...rest },
-}: {
-  node: Queries.MarkdownRemark
-}) => {
-  return {
-    objectID: id,
-    url: `https://til.swfz.io/${fields.slug}`,
-    github: `https://github.com/swfz/til/blob/master/content/blog/${fields.slug}index.md`,
-    text: rawMarkdownBody,
-    ...frontmatter,
-    ...fields,
-    ...rest,
-  }
-}
-
-const dataTransformer = ({ data }: Data) => {
-  console.log(`markdownRemark: ${data.postsRemark.edges.length} Records`)
-
-  return data.postsRemark.edges.map(pageToAlgoliaRecord)
-}
-
-export const queries = [
-  {
-    query: algoliaQuery,
-    transformer: dataTransformer,
-    indexName: process.env.ALGOLIA_INDEX_NAME, // overrides main index name, optional
-    settings: {
-      attributesToSnippet: [`text:50`],
-      attributesForFaceting: [`tags`],
-    },
-  },
-]
-
 // For gatsby-remark-related-posts
 export const remarkRelatedPostsOptions = {
   doc_lang: "ja", // optional
